+++ src/main/java/io/stargate/sgv2/jsonapi/service/schema/collections/CollectionRerankDef.java
@@ -104,12 +104,17 @@ public class CollectionRerankDef {
    * collections. Similarly, if the default provider doesn't have a default model, reranking will be
    * disabled.
    *
+   * @param isRerankingEnabledForAPI
    * @param rerankingProvidersConfig The configuration for all available reranking providers
    * @return A default-configured CollectionRerankDef
    */
   public static CollectionRerankDef configForNewCollections(
-      RerankingProvidersConfig rerankingProvidersConfig) {
+      boolean isRerankingEnabledForAPI, RerankingProvidersConfig rerankingProvidersConfig) {
     Objects.requireNonNull(rerankingProvidersConfig, "Reranking providers config cannot be null");
+    // If reranking is not enabled for the API, return disabled configuration
+    if (!isRerankingEnabledForAPI) {
+      return DISABLED;
+    }
     // Find the provider marked as default
     var defaultProviderEntry =
         rerankingProvidersConfig.providers().entrySet().stream()
@@ -235,7 +240,7 @@ public class CollectionRerankDef {
 
     // Case 1: No configuration provided - use defaults
     if (rerankingDesc == null) {
-      return configForNewCollections(providerConfigs);
+      return configForNewCollections(isRerankingEnabledForAPI, providerConfigs);
     }
 
     // Case 2: Validate 'enabled' flag is present
@@ -253,7 +258,7 @@ public class CollectionRerankDef {
     // Case 4: Enabled but no service config - use defaults
     var serviceConfig = rerankingDesc.rerankServiceDesc();
     if (serviceConfig == null) {
-      return configForNewCollections(providerConfigs);
+      return configForNewCollections(isRerankingEnabledForAPI, providerConfigs);
     }
 
     // Case 5: Full configuration - validate all components
