+++ src/test/java/io/stargate/sgv2/jsonapi/service/operation/collections/CreateCollectionOperationTest.java
@@ -35,7 +35,6 @@ import io.stargate.sgv2.jsonapi.service.cqldriver.CQLSessionCache;
 import io.stargate.sgv2.jsonapi.service.cqldriver.executor.QueryExecutor;
 import io.stargate.sgv2.jsonapi.service.reranking.configuration.RerankingProvidersConfig;
 import io.stargate.sgv2.jsonapi.service.schema.collections.CollectionLexicalConfig;
-import io.stargate.sgv2.jsonapi.service.schema.collections.CollectionRerankDef;
 import io.stargate.sgv2.jsonapi.service.testutil.MockAsyncResultSet;
 import io.stargate.sgv2.jsonapi.service.testutil.MockRow;
 import io.stargate.sgv2.jsonapi.testresource.NoGlobalResourcesTestProfile;
@@ -72,12 +71,8 @@ public class CreateCollectionOperationTest extends OperationTestBase {
     private final CollectionLexicalConfig LEXICAL_CONFIG =
         CollectionLexicalConfig.configForEnabledStandard();
 
-    private CollectionRerankDef RERANKING_CONFIG;
-
     @BeforeEach
-    public void init() {
-      RERANKING_CONFIG = CollectionRerankDef.configForNewCollections(rerankingProvidersConfig);
-    }
+    public void init() {}
 
     @Test
     public void createCollectionNoVector() {
