+++ src/test/java/io/stargate/sgv2/jsonapi/service/resolver/CreateCollectionCommandResolverTest.java
@@ -143,7 +143,7 @@ class CreateCollectionCommandResolverTest {
                             + "\"service\":{\"provider\":\"azureOpenAI\",\"modelName\":\"text-embedding-3-small\","
                             + "\"parameters\":{\"resourceName\":\"test\",\"deploymentId\":\"test\"}}},\"defaultId\":{\"type\":\"\"},"
                             + "\"lexical\":{\"enabled\":true,\"analyzer\":\"standard\"},"
-                            + "\"rerank\":{\"enabled\":true,\"service\":{\"provider\":\"nvidia\",\"modelName\":\"nvidia/llama-3.2-nv-rerankqa-1b-v2\",\"authentication\":null,\"parameters\":null}}}}"
+                            + "\"rerank\":{\"enabled\":false}}}"
                             + "}",
                         TableCommentConstants.SCHEMA_VERSION_VALUE);
               });
@@ -186,7 +186,7 @@ class CreateCollectionCommandResolverTest {
                         "{\"collection\":{\"name\":\"my_collection\",\"schema_version\":%s,\"options\":{\"indexing\":{\"deny\":[\"comment\"]},"
                             + "\"vector\":{\"dimension\":4,\"metric\":\"cosine\",\"sourceModel\":\"OTHER\"},\"defaultId\":{\"type\":\"\"},"
                             + "\"lexical\":{\"enabled\":true,\"analyzer\":\"standard\"},"
-                            + "\"rerank\":{\"enabled\":true,\"service\":{\"provider\":\"nvidia\",\"modelName\":\"nvidia/llama-3.2-nv-rerankqa-1b-v2\",\"authentication\":null,\"parameters\":null}}}}"
+                            + "\"rerank\":{\"enabled\":false}}}"
                             + "}",
                         TableCommentConstants.SCHEMA_VERSION_VALUE);
               });
